refactor: give MAIL FROM and RCPT TO a file each
protocol.go held the reader of the parameters of both commands, and every
extension that takes a parameter lands there: SIZE, BODY and AUTH of RFC
5321 and RFC 4954, RET and ENVID of RFC 3461, SMTPUTF8 of RFC 6531, and
NOTIFY and ORCPT on the recipient.

mail.go and rcpt.go take the reader of each command, with the handler
that calls it and the reply to a parameter that the server does not know.
The order is the one that the package keeps: handleDATA stands in
data.go, and handleMAIL stands in mail.go now.

protocol.go keeps the base of RFC 5321: the dispatcher, HELO and EHLO
with the keywords that EHLO answers, and RSET, NOOP and QUIT.

Nothing changes on the wire, and every declaration of the package is the
same one in another file.
